Separate chat requests from the TelegramClient
